# internal/server/worker/fiber/http/api/v1alpha/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Server API for MySQL auto failover. Generally must be called by CLI",
        "title": "MySQL auto failover server API",
        "contact": {
            "name": "Pavel Sapezhka",
            "url": "weastur.com",
            "email": "me@weastur.com"
        },
        "license": {
            "name": "Mozilla Public License Version 2.0",
            "url": "https://www.mozilla.org/en-US/MPL/2.0/"
        },
        "version": "v1alpha"
    },
    "host": "127.0.0.1:7080",
    "basePath": "/api/v1alpha",
    "paths": {
        "/raft/forget": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Forget the server. The server becomes non-voter and forgotten by the cluster in case of success\nand will not participate in the consensus",
                "tags": [
                    "raft"
                ],
                "summary": "Forget the server",
                "parameters": [
                    {
                        "description": "Forget request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RaftForgetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response with error details or success code",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        },
                        "headers": {
                            "X-API-Version": {
                                "type": "string",
                                "description": "API version, e.g. v1alpha"
                            },
                            "X-Ratelimit-Limit": {
                                "type": "int",
                                "description": "Rate limit value"
                            },
                            "X-Ratelimit-Remaining": {
                                "type": "int",
                                "description": "Rate limit remaining"
                            },
                            "X-Ratelimit-Reset": {
                                "type": "int",
                                "description": "Rate limit reset interval in seconds"
                            },
                            "X-Request-ID": {
                                "type": "string",
                                "description": "UUID of the request"
                            }
                        }
                    }
                }
            }
        },
        "/raft/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Return the raft cluster info with current server state and stats",
                "tags": [
                    "raft"
                ],
                "summary": "Return raft info",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Include extended stats",
                        "name": "include_stats",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Raft cluster info",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/RaftInfoResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-API-Version": {
                                "type": "string",
                                "description": "API version, e.g. v1alpha"
                            },
                            "X-Ratelimit-Limit": {
                                "type": "int",
                                "description": "Rate limit value"
                            },
                            "X-Ratelimit-Remaining": {
                                "type": "int",
                                "description": "Rate limit remaining"
                            },
                            "X-Ratelimit-Reset": {
                                "type": "int",
                                "description": "Rate limit reset interval in seconds"
                            },
                            "X-Request-ID": {
                                "type": "string",
                                "description": "UUID of the request"
                            }
                        }
                    }
                }
            }
        },
        "/raft/join": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Join the server to the cluster. The server becomes voter in case of success",
                "tags": [
                    "raft"
                ],
                "summary": "Join server to cluster",
                "parameters": [
                    {
                        "description": "Join request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RaftJoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response with error details or success code",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        },
                        "headers": {
                            "X-API-Version": {
                                "type": "string",
                                "description": "API version, e.g. v1alpha"
                            },
                            "X-Ratelimit-Limit": {
                                "type": "int",
                                "description": "Rate limit value"
                            },
                            "X-Ratelimit-Remaining": {
                                "type": "int",
                                "description": "Rate limit remaining"
                            },
                            "X-Ratelimit-Reset": {
                                "type": "int",
                                "description": "Rate limit reset interval in seconds"
                            },
                            "X-Request-ID": {
                                "type": "string",
                                "description": "UUID of the request"
                            }
                        }
                    }
                }
            }
        },
        "/raft/kv": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Set value in kv store for the key",
                "tags": [
                    "raft"
                ],
                "summary": "Set value for key",
                "parameters": [
                    {
                        "description": "KV set request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/KVSetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response with error details or success code",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        },
                        "headers": {
                            "X-API-Version": {
                                "type": "string",
                                "description": "API version, e.g. v1alpha"
                            },
                            "X-Ratelimit-Limit": {
                                "type": "int",
                                "description": "Rate limit value"
                            },
                            "X-Ratelimit-Remaining": {
                                "type": "int",
                                "description": "Rate limit remaining"
                            },
                            "X-Ratelimit-Reset": {
                                "type": "int",
                                "description": "Rate limit reset interval in seconds"
                            },
                            "X-Request-ID": {
                                "type": "string",
                                "description": "UUID of the request"
                            }
                        }
                    }
                }
            }
        },
        "/raft/kv/{key}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Return value for the key from kv store",
                "tags": [
                    "raft"
                ],
                "summary": "Return value of the key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key to receive value",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "KV get response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/KVGetResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-API-Version": {
                                "type": "string",
                                "description": "API version, e.g. v1alpha"
                            },
                            "X-Ratelimit-Limit": {
                                "type": "int",
                                "description": "Rate limit value"
                            },
                            "X-Ratelimit-Remaining": {
                                "type": "int",
                                "description": "Rate limit remaining"
                            },
                            "X-Ratelimit-Reset": {
                                "type": "int",
                                "description": "Rate limit reset interval in seconds"
                            },
                            "X-Request-ID": {
                                "type": "string",
                                "description": "UUID of the request"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete value in kv store for the key",
                "tags": [
                    "raft"
                ],
                "summary": "Delete value for key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key to delete value for",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response with error details or success code",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        },
                        "headers": {
                            "X-API-Version": {
                                "type": "string",
                                "description": "API version, e.g. v1alpha"
                            },
                            "X-Ratelimit-Limit": {
                                "type": "int",
                                "description": "Rate limit value"
                            },
                            "X-Ratelimit-Remaining": {
                                "type": "int",
                                "description": "Rate limit remaining"
                            },
                            "X-Ratelimit-Reset": {
                                "type": "int",
                                "description": "Rate limit reset interval in seconds"
                            },
                            "X-Request-ID": {
                                "type": "string",
                                "description": "UUID of the request"
                            }
                        }
                    }
                }
            }
        },
        "/version": {
            "get": {
                "description": "Return the version of running app. Not the API version, but the application",
                "tags": [
                    "aux"
                ],
                "summary": "Return version",
                "responses": {
                    "200": {
                        "description": "Version",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/VersionResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-API-Version": {
                                "type": "string",
                                "description": "API version, e.g. v1alpha"
                            },
                            "X-Ratelimit-Limit": {
                                "type": "int",
                                "description": "Rate limit value"
                            },
                            "X-Ratelimit-Remaining": {
                                "type": "int",
                                "description": "Rate limit remaining"
                            },
                            "X-Ratelimit-Reset": {
                                "type": "int",
                                "description": "Rate limit reset interval in seconds"
                            },
                            "X-Request-ID": {
                                "type": "string",
                                "description": "UUID of the request"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "KVGetResponse": {
            "description": "Response to the get request. Also contains 'exist' flag to distinguish between empty and non-existent string value",
            "type": "object",
            "properties": {
                "exist": {
                    "type": "boolean",
                    "example": true
                },
                "key": {
                    "type": "string",
                    "example": "key"
                },
                "value": {
                    "type": "string",
                    "example": "value"
                }
            }
        },
        "KVSetRequest": {
            "description": "Request to set the key-value pair",
            "type": "object",
            "properties": {
                "key": {
                    "type": "string",
                    "example": "key"
                },
                "value": {
                    "type": "string",
                    "example": "value"
                }
            }
        },
        "RaftForgetRequest": {
            "description": "Raft forget request with server metadata",
            "type": "object",
            "required": [
                "serverId"
            ],
            "properties": {
                "serverId": {
                    "type": "string",
                    "example": "maf-2"
                }
            }
        },
        "RaftInfoResponse": {
            "description": "Satatus of the raft cluster with servers metadata",
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string",
                    "example": "127.0.0.1:7081"
                },
                "id": {
                    "type": "string",
                    "example": "maf-1"
                },
                "servers": {
                    "description": "List of servers in the cluster",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RaftServer"
                    }
                },
                "state": {
                    "description": "State of the server in terms of the consensus: Leader, Follower, Candidate, etc.",
                    "type": "string",
                    "enum": [
                        "Follower",
                        "Candidate",
                        "Leader",
                        "Shutdown",
                        "Unknown"
                    ],
                    "example": "Leader"
                },
                "stats": {
                    "description": "Extended stats of the raft cluster",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "RaftJoinRequest": {
            "description": "Raft join request with server metadata",
            "type": "object",
            "required": [
                "addr",
                "serverId"
            ],
            "properties": {
                "addr": {
                    "type": "string",
                    "example": "10.1.2.3:7081"
                },
                "serverId": {
                    "type": "string",
                    "example": "maf-2"
                }
            }
        },
        "RaftServer": {
            "description": "Metadata of the server in the raft cluster",
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "127.0.0.1:7081"
                },
                "id": {
                    "type": "string",
                    "example": "maf-1"
                },
                "leader": {
                    "type": "boolean",
                    "example": true
                },
                "suffrage": {
                    "description": "Suffrage of the server in terms of the consensus: Voter, Nonvoter, Staging",
                    "type": "string",
                    "enum": [
                        "Voter",
                        "Nonvoter",
                        "Staging"
                    ],
                    "example": "Voter"
                }
            }
        },
        "Response": {
            "description": "Response wrapper to not build the API on top of outdated HTTP codes set",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Any structured data",
                    "type": "object"
                },
                "error": {
                    "description": "Error message. If status is not success, this field must be filled by a string with error message",
                    "type": "string",
                    "example": ""
                },
                "status": {
                    "description": "Response status\n* success - everything is OK\n* error   - something went wrong\n* warning - something went wrong, but it's not critical",
                    "type": "string",
                    "enum": [
                        "success",
                        "error",
                        "warning"
                    ],
                    "example": "success"
                }
            }
        },
        "VersionResponse": {
            "description": "Application version",
            "type": "object",
            "properties": {
                "version": {
                    "type": "string",
                    "example": "v1.0.0"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for the server. For now, only 'root' is allowed",
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    },
    "tags": [
        {
            "description": "Auxiliary endpoints",
            "name": "aux"
        },
        {
            "description": "Raft-related endpoints",
            "name": "raft"
        }
    ]
}